display finalized

+++ lib/Meter/Meter.cpp
@@ -9,7 +9,9 @@ Meter::Meter(Logger* logger)
     pinMode(Meter::ECHO_PIN, INPUT_PULLUP);
 }
 
-// returns the measured distance in meters
+/*
+ * returns the measured distance from the sensor to the column in meters
+ */
 float Meter::measure()
 {
     digitalWrite(Meter::TRIG_PIN, LOW);
@@ -20,7 +22,7 @@ float Meter::measure()
     digitalWrite(Meter::TRIG_PIN, LOW);
 
     // time in seconds
-    float timeTook = pulseIn(Meter::ECHO_PIN, HIGH, 100000) / 1000000;
+    float timeTook = (float) pulseIn(Meter::ECHO_PIN, HIGH, 100000) / 1000000;
     float distance = this->speedOfSound * timeTook;
 
     this->logger->info("Measurement taken. time: " + String(timeTook) + "s, distance: " + String(distance) + "m.");
